Refactor gql queries for meta-info
